requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
pycryptodome==3.20.0